Doc about configuration in Cargo.toml has a typo
I discovered this when trying to configure this in my `Cargo.toml`.
I prefer to have the configuration for this utility in there, instead of `release.toml` and was glad it's an option, but when I was trying to use `metadata.release`. Turns out it's `package.metadata.release`.

I hope this helps others with the same issue I encountered.